Added a possible CSV format. Need to investigate i/o numbers.
Man pages and other sources remark that the i/o numbers are in "blocks", which
are probably 512 bytes (at least, they used to be that size).  The same sources
say that these numbers may be misleading because they are "real" i/o and will
not reflect cache operations.  On macos 14.4.1, the i/o block operations numbers
are always both zero.  Is that because macos, Darwin, and/or BSD kernels do not
populate those fields?  Or perhaps all disk i/o (which is what I was testing) is
cached and it will never show up?  Or maybe operations on SSD devices do not
count?  More investigation is needed, to determine if the i/o numbers are useful
under any circumstances, and on any platforms.