Implemented experimental report with box plots
The report is currently obtained by using the 'reduce' command, e.g.

./bestguess reduce -f raw.csv

Where the specified file is the raw data output (generated by running an
experiment using `./bestguess -o raw.csv`.

The functionality of the 'reduce' command will change, but for now it is the
only way to access a full descriptive summary of the data distribution.