Narrowing down the set of applicable statistics relating two samples
Specifically: Each sample is a collection of performance observations, and we
have demonstrated that the sample fails a test for normality.  We want to
compare the medians of the two samples, e.g. to know with what confidence that
the median run time for one command is really less than the median run time for
another.