Can replace Z table with iterative calculation if we want
See cdf.c.  We would then be able to work with Z scores up through 7 or 8, which
would let us calculate AD scores in more situations.  Also, we can state clearly
that, e.g. a Z score of 7 indicates a one in 390 BILLION chance of seeing that
observation in a normally distributed sample.