Refactored to use a Ranking struct.  See TODO items below.
TODO 1: Fix CLI so that if config string is omitted, user sees a more
informative error message.  Test using:

    ./bestreport -x -R none test/rawrosie.csv

TODO 2: Check the output, as we are now producing it using a different approach.
Make sure the rankings are correct and the explanations as well.

TODO 3: Decide on a format for the explanations.

TODO 4: Get rid of print_summary() and replace with print_ranking().