Making minor changes in preparation for some refactoring.
Want to give useful and consistent names to the reports and to the pieces of
reports (like command "announcements", which is a good one).  Would like to
separate execution from reporting, but we get a better UX if we announce each
command before running it and report on it before moving on to the next command.

So we will need to interleave execution of an experiment with reporting.  But
BestGuess has the ability to reproduce the same reports later on --
post-execution -- using saved raw data.  For that use case, we want to display
those same reports, requiring APIs to do just that.